Moved event into correct spot
